check arg

+++ stopRPi.pl
@@ -7,7 +7,7 @@ my @pidList;
 my @ProcList = ('SystemMonitor.pl','MapNetwork.pl','WatchDog.pl','ConfigServer.pl','ShutdownNUC.pl','i2c_apc_v6.pl',
 				'RemoteLog.pl', 'RunNVR.pl', 'RemoteData.pl','WatchSSD.pl','i2c_apc_v6.pl','Watcher.pl','WatchRPi.pl');
 my $ret;
-if ($ARGV[0] eq 'kill') {
+if (( defined($ARGV[0])) && ($ARGV[0] eq 'kill')) {
 	my $ret = `ps -elf |grep perl` ;
 	print $ret;
 } else {
@@ -24,7 +24,7 @@ foreach my $curProc (@ProcList) {
  	}
 	push(@pidList,$pid) unless $pid == -1;
 }
-if ( (scalar @pidList > 0) && ($ARGV[0] eq 'kill')) {
+if ( (scalar @pidList > 0) && defined($ARGV[0]) && ($ARGV[0] eq 'kill')) {
 	my $cmd = "kill " . join(" ",@pidList);
 	`$cmd`;	
 	$ret = `ps -elf |grep perl`;
